+++ axlearn/common/config.py
@@ -10,7 +10,8 @@
 
 Adapted from https://github.com/tensorflow/lingvo/blob/master/lingvo/core/hyperparams.py.
 
-See https://tinyurl.com/ml-api-styles for the design principles behind this config library.
+See https://github.com/apple/axlearn/blob/main/docs/ml_api_style.md
+for the design principles behind this config library.
 
 Example usage for configuring a module: