+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changes
     * Upgrade Python to 3.10
+    * Fall back to triton backend for qkv in fp32 or with bias on gpu flash attention.
 
 ## 0.1.1